refactor(agent): split agent loop phases
Constraint: Preserve the public loop API, event and trace ordering, continuation state, tool admission, and replay safety while removing the orchestration monolith.
Rejected: Renaming the 2,338-line loop or moving it unchanged into another large file.
Confidence: high
Scope-risk: broad
Directive: Keep agent-loop as the lifecycle orchestrator and isolate provider, tool, budget, admission, execution, metrics, and public type ownership.
Tested: Final pnpm typecheck, 112 focused tests, pnpm build, and packaged Agent session; full logic (1,973), regressions (673), Scenario Lab, worker recovery, and extension smoke before the final pure metrics relocation.
Not-tested: Live provider credentials; release evidence regeneration; Chrome Web Store upload, review, or publication.
